% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lot_map_to_ots.R
\name{.compute_ots}
\alias{.compute_ots}
\title{Internal OTS computation}
\usage{
.compute_ots(deltas, thresholds)
}
\arguments{
\item{deltas}{A data frame or `DataFrame` containing molecular-change
estimates.}

\item{thresholds}{A data frame or `DataFrame` containing feature-specific
thresholds.}
}
\value{
A `DataFrame` containing Omics Trajectory Signature assignments.
}
\description{
Converts molecular deltas into transition states using feature-specific
thresholds and combines those states into Omics Trajectory Signature
labels.
}
\keyword{internal}
