% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{package}
\name{LongOmicsTraj-package}
\alias{LongOmicsTraj}
\alias{LongOmicsTraj-package}
\title{LongOmicsTraj package imports}
\description{
Provides methods for estimating, classifying and comparing longitudinal molecular trajectories across repeated study visits using empirical, mixed-effects, generalized additive model and polynomial regression estimators. Consecutive molecular changes are converted into interpretable Omics Trajectory Signatures (OTS) describing stable, sustained, transient and rebound temporal behaviours, enabling topology- aware clustering and consensus trajectory inference. LongOmicsTraj further provides a flexible query framework for directly retrieving molecular features satisfying simple or complex trajectory definitions, including multi-transition patterns, logical combinations of trajectory states, group-specific behaviours, and comparisons of trajectory topologies across treatment arms, experimental conditions or other longitudinal study arms.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nzkermani/LongOmicsTraj}
  \item Report bugs at \url{https://github.com/nzkermani/LongOmicsTraj/issues}
}

}
\author{
\strong{Maintainer}: Nazanin Zounemat-Kermani \email{nonakermani@gmail.com} (\href{https://orcid.org/0000-0003-2479-3861}{ORCID})

Authors:
\itemize{
  \item Nazanin Zounemat-Kermani \email{nonakermani@gmail.com} (\href{https://orcid.org/0000-0003-2479-3861}{ORCID})
}

}
\keyword{internal}
