% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-LongOmicsTraj.R
\name{LongOmicsTraj}
\alias{LongOmicsTraj}
\title{Construct a LongOmicsTraj object}
\usage{
LongOmicsTraj(
  mae = MultiAssayExperiment::MultiAssayExperiment(),
  vmm = S4Vectors::DataFrame(),
  deltas = S4Vectors::DataFrame(),
  thresholds = S4Vectors::DataFrame(),
  ots = S4Vectors::DataFrame(),
  relationships = S4Vectors::DataFrame(),
  results = list(),
  cache = list(),
  settings = list(),
  history = list()
)
}
\description{
Construct a LongOmicsTraj object
}
