% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-deltas.R
\name{lot_compute_deltas,LongOmicsTraj-method}
\alias{lot_compute_deltas,LongOmicsTraj-method}
\title{Compute deltas between visits}
\usage{
\S4method{lot_compute_deltas}{LongOmicsTraj}(object, assay, visits, overwrite = FALSE)
}
\description{
Compute deltas between visits
}
