% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lot_glucold_expression}
\alias{lot_glucold_expression}
\title{GLUCOLD expression matrix}
\format{
A numeric matrix with molecular features in rows and samples
in columns.
}
\source{
NCBI Gene Expression Omnibus accession GSE36221.
}
\usage{
data(lot_glucold_expression)
}
\description{
Pre-processed longitudinal transcriptomic measurements for 71 selected
candidate and hypothesis-driven genes from the GLUCOLD study.
}
\keyword{datasets}
