% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{NOT}
\alias{NOT}
\title{Negate a topology query condition}
\usage{
NOT(x)
}
\arguments{
\item{x}{Character vector of topology labels to exclude.}
}
\value{
A query condition used by lot_query().
}
\description{
Negate a topology query condition
}
