% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cluster_spaghetti_single.R
\name{plot_cluster_spaghetti_single}
\alias{plot_cluster_spaghetti_single}
\title{Plot one clustered molecular-feature trajectory profile}
\usage{
plot_cluster_spaghetti_single(object, cluster_id, assay_name, visits_vector)
}
\arguments{
\item{object}{A `LongOmicsTraj` object containing cluster relationships,
gene-level OTS results, cluster trajectories and OTS metrics.}

\item{cluster_id}{Identifier of the cluster to display.}

\item{assay_name}{Name of the assay to display.}

\item{visits_vector}{Character vector giving visits in chronological order.}
}
\value{
A `ggplot` object.
}
\description{
Creates a spaghetti plot for a single molecular-feature cluster.
Individual feature trajectories are coloured by their Omics Trajectory
Signature, while the representative cluster trajectory is displayed in
black.
}
