% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lot_summary.R
\name{print.lot_summary}
\alias{print.lot_summary}
\title{Print a LongOmicsTraj summary}
\usage{
\method{print}{lot_summary}(x, ...)
}
\arguments{
\item{x}{A `lot_summary` object.}

\item{...}{Additional arguments, currently ignored.}
}
\value{
The summary object, returned invisibly.
}
\description{
Print a LongOmicsTraj summary
}
