% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-LongOmicsTraj.R
\name{show,LongOmicsTraj-method}
\alias{show,LongOmicsTraj-method}
\title{Display a LongOmicsTraj object}
\usage{
\S4method{show}{LongOmicsTraj}(object)
}
\arguments{
\item{object}{A `LongOmicsTraj` object.}
}
\value{
The object is returned invisibly.
}
\description{
Prints a concise summary of the assays, study design and analysis
results stored in a `LongOmicsTraj` object.
}
\examples{
data("lot_glucold")

## Printing the object invokes the show() method.
lot_glucold

## The object is returned invisibly.
invisible(lot_glucold)

}
